domain_service/app/requirements.txt
python-json-logger
python-logstash
fastapi
uvicorn
grpcio
grpcio-tools
pymongo
pika